requirements.txt
asn1crypto==0.22.0
astroid==1.5.3
backports.functools-lru-cache==1.4
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.0.3
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Flask==0.12.2
Flask-API==1.0
Flask-Cors==3.0.3
flask-marshmallow==0.8.0
Flask-Script==2.0.5
Flask-SocketIO==2.9.2
geoip2==2.5.0
greenlet==0.4.12
hpfeeds==1.0
idna==2.6
ipaddress==1.0.18
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.13.6
maxminddb==1.3.0
mccabe==0.6.1
pycparser==2.18
pylint==1.7.2
pyOpenSSL==17.2.0
python-engineio==1.7.0
python-socketio==1.8.0
PyYAML==3.12
requests==2.18.4
singledispatch==3.4.0.3
six==1.10.0
socketIO-client==0.7.2
urllib3==1.22
websocket-client==0.44.0
Werkzeug==0.12.2
wrapt==1.10.11
